Split VectorSet* code into partial files
